pytest/requirements.txt
# Dependências para testes de failover e medição de RTO

# Framework de testes
pytest>=7.4.0
pytest-timeout>=2.1.0
pytest-html>=3.2.0

# Conexão PostgreSQL
psycopg2-binary>=2.9.7

# Manipulação de dados
dataclasses-json>=0.6.0

# Utilitários
python-dateutil>=2.8.2
